+++ src/main/java/chess/controller/SpringChessController.java
@@ -1,8 +1,8 @@
 package chess.controller;
 
 import chess.service.SpringChessService;
-import chess.viewdto.ChessGameDto;
-import chess.webdto.MoveRequestDto;
+import chess.webdto.view.ChessGameDto;
+import chess.webdto.view.MoveRequestDto;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
