+++ src/main/java/chess/webdto/WordConstants.java
@@ -4,4 +4,10 @@ public class WordConstants {
     public static final String WHITE = "white";
     public static final String BLACK = "black";
 
+    public static String convert(boolean whiteTeamTurn) {
+        if(whiteTeamTurn){
+            return WordConstants.WHITE;
+        }
+        return WordConstants.BLACK;
+    }
 }
